+++ src/Engine/Objects/World/World.cpp
@@ -2,7 +2,7 @@
 #include "Rendering/Buffers/UniformBufferFactory.h"
 #include <glad/glad.h>
 #include <GLFW/glfw3.h>
-void World::Tick(const float& InDeltaTime)
+void World::Tick(const double& InDeltaTime)
 {
 	Object::Tick(DeltaTime);
